organise the code a bit, prepare for types
